+++ CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - LuckPerms permission events
   - `addPermission` event adds a permission to a player
   - `removePermission` event removes a permission from a player
+- Location Vectors can now be chained
 ### Changed
 - `AureliumSkills` was renamed to `AuraSkills` the following was changed in BQ
   - `aureliumskillslevel` condition was renamed to `auraskillslevel`